tests/wpt/web-platform-tests/webvtt/parsing/cue-text-parsing/dat/tree-building.dat

#data
test
#errors
#document-fragment
| "test"

#data
<c>test
#errors
#document-fragment
| <span>
|   "test"

#data
<i>test
#errors
#document-fragment
| <i>
|   "test"

#data
<b>test
#errors
#document-fragment
| <b>
|   "test"

#data
<u>test
#errors
#document-fragment
| <u>
|   "test"

#data
<ruby>test
#errors
#document-fragment
| <ruby>
|   "test"

#data
<rt>test
#errors
#document-fragment
| "test"

#data
<ruby>test<rt>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     "test"

#data
<ruby>test<rt>test</rt>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     "test"
|   "test"

#data
<ruby>test<rt>test</rt></ruby>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     "test"
| "test"

#data
<ruby>test<rt>test</ruby>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     "test"
| "test"

#data
<ruby>test<rt><b>test</rt></ruby>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     <b>
|       "test"
|       "test"

#data
<ruby>test<rt><b>test</ruby>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     <b>
|       "test"
|       "test"

#data
<ruby>test<rt><b>test</rt></ruby></b>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     <b>
|       "test"
|     "test"

#data
<ruby>test<rt><b>test</rt></b></ruby>test
#errors
#document-fragment
| <ruby>
|   "test"
|   <rt>
|     <b>
|       "test"
| "test"

#data
<v>test
#errors
#document-fragment
| <span>
|   title=""
|   "test"

#data
<v a>test
#errors
#document-fragment
| <span>
|   title="a"
|   "test"

#data
<v a b>test
#errors
#document-fragment
| <span>
|   title="a b"
|   "test"

#data
<v.a>test
#errors
#document-fragment
| <span>
|   class="a"
|   title=""
|   "test"

#data
<v.a.b>test
#errors
#document-fragment
| <span>
|   class="a b"
|   title=""
|   "test"
